Conflicts:
	src/man/Man.cpp
	src/man/corpus/Sensors.cpp
	src/man/include/Profiler.cpp
	src/man/vision/CMakeLists.txt
	src/man/vision/EdgeDetector.cpp
	src/man/vision/EdgeDetector.h
	src/man/vision/Field.h
	src/man/vision/FieldLines.cpp
	src/man/vision/FieldLinesDetector.cpp
	src/man/vision/HoughSpace.cpp
	src/man/vision/HoughSpace.h
	src/man/vision/Profiler.h
	src/man/vision/Threshold.cpp
	src/man/vision/Vision.cpp
	src/man/vision/Vision.h
	src/man/vision/VisualLine.cpp
	src/man/vision/VisualLine.h
	src/man/vision/offline/OfflineVision.cpp
	src/man/vision/test/EdgeDetectorTest.cpp
	src/man/vision/test/HoughSpaceTest.cpp
	src/tool/TOOL/Learning/VisionState.java
	src/tool/TOOL/Vision/TOOLVisionLink.cpp

+++ src/man/include/NBMatrixMath.cpp
@@ -76,5 +76,5 @@ const ufvector3 NBMath::solve(ufmatrix3 &A,
 
 const float NBMath::length(ufvector4 &v) {
 
-    return sqrt(v(0) * v(0) + v(1) * v(1) + v(2) * v(2));
+    return sqrtf(v(0) * v(0) + v(1) * v(1) + v(2) * v(2));
 }
